Removed sequence protocol, move to test code
