+++ composer.json
@@ -19,9 +19,6 @@
         "psr-0": {
             "Immutable\\": "./src",
             "Functional\\": "./src"
-        },
-        "files": [
-            "src/protocols/all.php"
-        ]
+        }
     }
 }
