--- src/Functional/Protocol/SequenceInterface.php
@@ -1,9 +1,0 @@
-<?php
-namespace Functional\Protocol;
-
-interface SequenceInterface
-{
-    public static function first($coll);
-    
-    public static function rest($coll);
-}
