--- src/protocols/all.php
@@ -1,1 +1,0 @@
-<?php
